tests/submission/W3C/domhtml/HTMLObjectElement19.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-object-codetype" rel="help">
<link rel="author" title="NIST">
</head>
<body>

<object codetype="text/plain" data="/support/html/sample.html" height="80" type="text/html" width="550"></object>

<script>
var domtest = async_test("The codeType IDL attribute of the object element must reflect the element's codetype content attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcodetype;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vcodetype = testNode.codeType;

      assert_equals(vcodetype,"text/plain");
       
    domtest.done();
}));

</script>
<p>The codeType IDL attribute of the object element must reflect the element's codetype content attribute.</p>
<div id="log"></div>
</body></html>
